requirements.txt
pytorch-lightning==0.6.0
pyyaml==5.3
torchvision==0.5.0
